add: dashboard and router
