+++ src/app/components/Navbar/Navbar.module.css
@@ -1,15 +1,3 @@
-
-.Navbar {
-    width: 100%;
-    height: 70px;
-    color: black;
-    background-color: white;
-    box-shadow: 2px 2px 10px #6177f0;
-    z-index: 2;
-    position: absolute;
-    border-radius: 0px 0px 0px 50px;
-}
-
 .Navbar {
     width: 100%;
     height: 70px;
@@ -19,4 +7,5 @@
     z-index: 2;
     position: absolute;
     border-radius: 0px 0px 0px 50px;
+    top: 0;
 }
