+++ src/app/layout.tsx
@@ -5,6 +5,7 @@ import "./globals.css";
 import "./styles/inputs.css";
 import "./styles/buttons.css";
 import "./styles/cards.css";
+import "./styles/manual.css";
 
 const geistSans = localFont({
   src: "./fonts/GeistVF.woff",
