+++ src/app/pages/dashboard/dashboard.page.css
@@ -1,0 +1,3 @@
+.dashboard-page {
+    height: 100vh;
+}
