+++ src/app/styles/cards.css
@@ -39,3 +39,11 @@
     border-color: blue;
     border-width: 10px;
   }
+
+  .dashboard-card {
+    background-color: white;
+    box-shadow: 2px 2px 10px #34353a;
+    margin-top: 7rem;
+    width: 80%;
+    border-radius: 10px;
+  }
